requirements.txt
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
fastapi==0.110.0
h11==0.14.0
httpcore==1.0.4
httptools==0.6.1
httpx==0.27.0
idna==3.6
outcome==1.3.0.post0
packaging==23.2
pydantic==2.6.3
pydantic_core==2.16.3
PySocks==1.7.1
python-dotenv==1.0.1
python-telegram-bot==21.0.1
PyYAML==6.0.1
requests==2.31.0
selenium==4.18.1
setuptools==69.1.1
sniffio==1.3.1
sortedcontainers==2.4.0
starlette==0.36.3
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.10.0
urllib3==2.2.1
uvicorn==0.28.0
uvloop==0.19.0
watchfiles==0.21.0
webdriver-manager==4.0.1
websockets==12.0
wheel==0.42.0
wsproto==1.2.0
